+++ src/com/amplitude/api/AmplitudeClient.java
@@ -271,9 +271,19 @@ public class AmplitudeClient {
         logEvent(eventType, eventProperties, false);
     }
 
+    public void logEvent(String eventType, JSONObject eventProperties, JSONObject groups) {
+        logEvent(eventType, eventProperties, groups, false);
+    }
+
     public void logEvent(String eventType, JSONObject eventProperties, boolean outOfSession) {
+        logEvent(eventType, eventProperties, null, outOfSession);
+    }
+
+    public void logEvent(String eventType, JSONObject eventProperties, JSONObject groups, boolean outOfSession) {
         if (validateLogEvent(eventType)) {
-            logEventAsync(eventType, eventProperties, null, null, getCurrentTimeMillis(), outOfSession);
+            logEventAsync(
+                eventType, eventProperties, null, null, groups, getCurrentTimeMillis(), outOfSession
+            );
         }
     }
 
@@ -282,8 +292,14 @@ public class AmplitudeClient {
     }
 
     public void logEventSync(String eventType, JSONObject eventProperties, boolean outOfSession) {
+        logEventSync(eventType, eventProperties, null, outOfSession);
+    }
+
+    public void logEventSync(String eventType, JSONObject eventProperties, JSONObject groups, boolean outOfSession) {
         if (validateLogEvent(eventType)) {
-            logEvent(eventType, eventProperties, null, null, getCurrentTimeMillis(), outOfSession);
+            logEvent(
+                eventType, eventProperties, null, null, groups, getCurrentTimeMillis(), outOfSession
+            );
         }
     }
 
@@ -302,7 +318,7 @@ public class AmplitudeClient {
 
     protected void logEventAsync(final String eventType, JSONObject eventProperties,
             final JSONObject apiProperties, final JSONObject userProperties,
-            final long timestamp, final boolean outOfSession) {
+            JSONObject groups, final long timestamp, final boolean outOfSession) {
         // Clone the incoming eventProperties object before sending over
         // to the log thread. Helps avoid ConcurrentModificationException
         // if the caller starts mutating the object they passed in.
@@ -312,17 +328,25 @@ public class AmplitudeClient {
             eventProperties = cloneJSONObject(eventProperties);
         }
 
+        if (groups != null) {
+            groups = cloneJSONObject(groups);
+        }
+
         final JSONObject copyEventProperties = eventProperties;
+        final JSONObject copyGroups = groups;
         runOnLogThread(new Runnable() {
             @Override
             public void run() {
-                logEvent(eventType, copyEventProperties, apiProperties, userProperties, timestamp, outOfSession);
+                logEvent(
+                    eventType, copyEventProperties, apiProperties,
+                    userProperties, copyGroups, timestamp, outOfSession
+                );
             }
         });
     }
 
     protected long logEvent(String eventType, JSONObject eventProperties, JSONObject apiProperties,
-            JSONObject userProperties, long timestamp, boolean outOfSession) {
+            JSONObject userProperties, JSONObject groups, long timestamp, boolean outOfSession) {
         logger.d(TAG, "Logged event to Amplitude: " + eventType);
 
         if (optOut) {
@@ -386,6 +410,7 @@ public class AmplitudeClient {
                     : truncate(eventProperties));
             event.put("user_properties", (userProperties == null) ? new JSONObject()
                     : truncate(userProperties));
+            event.put("groups", (groups == null) ? new JSONObject() : truncate(groups));
         } catch (JSONException e) {
             logger.e(TAG, e.toString());
         }
@@ -573,7 +598,7 @@ public class AmplitudeClient {
         }
 
         long timestamp = getLastEventTime();
-        logEvent(sessionEvent, null, apiProperties, null, timestamp, false);
+        logEvent(sessionEvent, null, apiProperties, null, null, timestamp, false);
     }
 
     void onExitForeground(final long timestamp) {
@@ -624,7 +649,9 @@ public class AmplitudeClient {
         } catch (JSONException e) {
         }
 
-        logEventAsync(REVENUE_EVENT, null, apiProperties, null, getCurrentTimeMillis(), false);
+        logEventAsync(
+                REVENUE_EVENT, null, apiProperties, null, null, getCurrentTimeMillis(), false
+        );
     }
 
     // maintain for backwards compatibility
@@ -676,7 +703,23 @@ public class AmplitudeClient {
             return;
         }
         logEventAsync(Constants.IDENTIFY_EVENT, null, null,
-                identify.userPropertiesOperations, getCurrentTimeMillis(), false);
+                identify.userPropertiesOperations, null, getCurrentTimeMillis(), false);
+    }
+
+    public void setGroup(String groupType, Object groupName) {
+        if (TextUtils.isEmpty(groupType) || !contextAndApiKeySet("setGroup()")) {
+            return;
+        }
+        JSONObject group = null;
+        try {
+            group = new JSONObject().put(groupType, groupName);
+        }
+        catch (JSONException e) {
+            logger.e(TAG, e.toString());
+        }
+        Identify identify = new Identify().setUserProperty(groupType, groupName);
+        logEventAsync(Constants.IDENTIFY_EVENT, null, null, identify.userPropertiesOperations,
+                group, getCurrentTimeMillis(), false);
     }
 
     public JSONObject truncate(JSONObject object) {