+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Update to OKHttp v3.0.1.
 * Add support for prepend user property operation.
+* Fix bug where merging events for upload causes array index out of bounds exception.
 * Migrate shared preferences (userId and event meta data) to Sqlite db to support apps with multiple processes.
 
 ## 2.5.1 (March 14, 2016)